🌈 style: 减少编译警告

+++ src/slic3r/GUI/Widgets/StateColor.hpp
@@ -61,7 +61,7 @@ public:
     void clear();
 
 public:
-    int count() const { return statesList_.size(); }
+    int count() const { return static_cast<int>(statesList_.size()); }
 
     int states() const;
